Bot/requirements.txt
discord.py==2.4.0
dataclasses-json==0.6.7
PyNaCl==1.5.0